backend/pyproject.toml
[tool.poetry]
name = "immo-backend"
version = "0.1.0"
description = "Immo backend application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.110.0"
uvicorn = "0.27.1"
pydantic = "2.6.1"
pydantic-settings = "2.2.1"
python-dotenv = "1.0.1"
anthropic = "0.18.1"
httpx = "0.26.0"
python-jose = "3.3.0"
passlib = "1.7.4"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.1"
tenacity = "^8.2.3"
openai = "^1.12.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
